// extension/skills/noggin/noggin-api.d.ts
export type ItemKey = string;
export type ItemPath = string;
export type IsoTimestamp = string;

export interface Note {
  timestamp: IsoTimestamp;
  text: string;
}

export interface Item {
  key: ItemKey;
  parentKey: ItemKey | null;
  title: string;
  done: boolean;
  pushedAt?: IsoTimestamp;
  closedAt?: IsoTimestamp | null;
  notes: Note[];
}